requirements.txt
python-dotenv

websockets
requests
fastapi
streamlit
streamlit_authenticator

uvicorn
#qrcode
